Moved search methods to a new class
