+++ src/test/java/se/qxx/protodb/test/TestSearchRepeated.java
@@ -14,6 +14,7 @@ import com.google.protobuf.ByteString;
 import se.qxx.protodb.JoinResult;
 import se.qxx.protodb.ProtoDB;
 import se.qxx.protodb.ProtoDBScanner;
+import se.qxx.protodb.Searcher;
 import se.qxx.protodb.exceptions.IDFieldNotFoundException;
 import se.qxx.protodb.exceptions.SearchFieldNotFoundException;
 import se.qxx.protodb.test.TestDomain.ObjectTwo;
@@ -103,7 +104,7 @@ public class TestSearchRepeated {
 					.build();
 			
 			ProtoDBScanner scanner = new ProtoDBScanner(o1);
-			JoinResult result = scanner.getJoinQuery(false);
+			JoinResult result = Searcher.getJoinQuery(scanner, false);
 
 			String expected = "SELECT "
 					+ "A.[ID] AS A_ID, "
